refactor: replace `int` with `byte` for handler type and update serialization logic
- Change handler type declaration from `int` to `byte` for memory optimization.
- Adjust serialization/deserialization to reflect the handler type size change.
- Refactor related constants, comments, and buffer operations for consistency.

+++ core/src/main/java/cn/ttplatform/wh/support/DefaultMessageCodec.java
@@ -1,16 +1,16 @@
 package cn.ttplatform.wh.support;
 
-import java.util.List;
-
 import cn.ttplatform.wh.cmd.RequestFailedCommand;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.ByteToMessageCodec;
 import lombok.extern.slf4j.Slf4j;
 
+import java.util.List;
+
 @Slf4j
 public class DefaultMessageCodec extends ByteToMessageCodec<Distributable>{
-    private static final int FIXED_MESSAGE_HEADER_LENGTH = Integer.BYTES + Integer.BYTES;
+    private static final int FIXED_MESSAGE_HEADER_LENGTH = Byte.BYTES + Integer.BYTES;
 
     private final DistributableSerializerRegistry serializerRegistry;
     private RequestFailedCommand requestFailedCommand;
@@ -30,9 +30,9 @@ public class DefaultMessageCodec extends ByteToMessageCodec<Distributable>{
         if (in.readableBytes() < FIXED_MESSAGE_HEADER_LENGTH) {
             return;
         }
-        // 4(type) + 4(contentLength) + byte[contentLength]
+        // 1(type) + 4(contentLength) + byte[contentLength]
         in.markReaderIndex();
-        int type = in.readInt();
+        byte type = in.readByte();
         int contentLength = in.readInt();
         int newReaderIndex = in.readerIndex() + contentLength;
         if (in.readableBytes() < contentLength) {
