+++ server/src/main/java/cn/ttplatform/wh/message/handler/PreVoteResultMessageHandler.java
@@ -1,11 +1,11 @@
 package cn.ttplatform.wh.message.handler;
 
-import cn.ttplatform.wh.message.PreVoteResultMessage;
-import cn.ttplatform.wh.constant.DistributableType;
 import cn.ttplatform.wh.GlobalContext;
 import cn.ttplatform.wh.Node;
+import cn.ttplatform.wh.constant.DistributableType;
 import cn.ttplatform.wh.group.Cluster;
 import cn.ttplatform.wh.group.Phase;
+import cn.ttplatform.wh.message.PreVoteResultMessage;
 import cn.ttplatform.wh.role.Follower;
 import cn.ttplatform.wh.support.AbstractDistributableHandler;
 import cn.ttplatform.wh.support.Distributable;
@@ -24,7 +24,7 @@ public class PreVoteResultMessageHandler extends AbstractDistributableHandler {
     }
 
     @Override
-    public int getHandlerType() {
+    public byte getHandlerType() {
         return DistributableType.PRE_VOTE_RESULT;
     }
